requirements.txt
fastapi
httpx
loguru
pydantic
pyexecjs
orjson
jsonpath
uvicorn[standard]
fastapi[all]
pyyaml